+++ journalkeeper-core/src/main/java/com/jd/journalkeeper/core/journal/NosuchPartitionException.java
@@ -5,13 +5,13 @@ package com.jd.journalkeeper.core.journal;
  * Date: 2019-04-25
  */
 public class NosuchPartitionException extends RuntimeException {
-    private final short partition;
-    public NosuchPartitionException(short partition) {
+    private final int partition;
+    public NosuchPartitionException(int partition) {
         super("No such partition: " + partition + "!");
         this.partition = partition;
     }
 
-    public short getPartition() {
+    public int getPartition() {
         return partition;
     }
 }
